requirements.txt
matplotlib
scipy
tqdm
opencv-python
numpy
scikit-image
torchvision
timm
git+https://github.com/ChaoningZhang/MobileSAM.git
rembg
scikit-learn